+++ Sources/KeyboardKit/Input/StandardKeyboardInputSetProvider.swift
@@ -11,12 +11,14 @@ import Foundation
 /**
  This standard keyboard inputset provider is used by default
  by KeyboardKit, and provides the standard input set for the
- current locale, if any. You can inherit and customize it to
- create your own provider that builds on this foundation.
+ current locale, if any.
  
  The dictionaries contain the currently supported input sets.
  They are not part of the protocol, but you can use them and
  extend them in your own subclasses.
+ 
+ You can inherit and customize this class to create your own
+ provider that builds on this foundation.
  */
 open class StandardKeyboardInputSetProvider: KeyboardInputSetProvider {
     
@@ -24,7 +26,7 @@ open class StandardKeyboardInputSetProvider: KeyboardInputSetProvider {
         self.locale = locale
     }
     
-    
+
     private let locale: Locale