+++ Sources/KeyboardKit/Input/StaticKeyboardInputSetProvider.swift
@@ -9,14 +9,11 @@
 import Foundation
 
 /**
- This standard keyboard inputset provider is used by default
- by KeyboardKit, and provides the standard input set for the
- current locale, if any. You can inherit and customize it to
- create your own provider that builds on this foundation.
+ This keyboard input set provider returns the input set that
+ it is initialized with.
  
- The dictionaries contain the currently supported input sets.
- They are not part of the protocol, but you can use them and
- extend them in your own subclasses.
+ It can be used to return a static input set, without taking
+ factors like locales into consideration.
  */
 public class StaticKeyboardInputSetProvider: KeyboardInputSetProvider {